chore: harden backend polish audit findings

+++ src/schemas/admin.ts
@@ -3,7 +3,7 @@ import z from 'zod';
 
 export const paginationQuerySchema = z.object({
   cursor: z.string().optional(),
-  limit: z.coerce.number().min(1).max(50).optional(),
+  limit: z.coerce.number().int().min(1).max(50).optional(),
 });
 
 const normalizeUuidListQueryParam = (value: unknown) => {
